RAG/requirements2.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.27
botocore==1.38.27
browser-use==0.1.48
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
courlan==1.3.2
Cython==3.1.1
dataclasses-json==0.6.7
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
gradio==5.27.0
gradio_client==1.9.0
greenlet==3.2.2
groovy==0.1.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
hyperframe==6.1.0
ibm-cos-sdk==2.13.5
ibm-cos-sdk-core==2.13.5
ibm-cos-sdk-s3transfer==2.13.5
ibm_watsonx_ai==1.3.23
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json_repair==0.46.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jusText==3.0.2
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-community==0.3.20
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ibm==0.3.10
langchain-mcp-adapters==0.0.9
langchain-mistralai==0.2.4
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.34
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.43
lomond==0.3.3
lxml==5.4.0
lxml_html_clean==0.4.2
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mem0ai==0.1.93
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
ollama==0.5.1
openai==1.82.1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyMuPDF==1.25.5
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruff==0.11.12
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
screeninfo==0.8.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tld==0.13.1
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.1.0
tornado==6.5.1
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.40.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.1.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
